+++ src/app/inventory/d2-stores.ts
@@ -7,11 +7,13 @@ import { currentAccountSelector } from 'app/accounts/selectors';
 import { t } from 'app/i18next-t';
 import { maxLightItemSet } from 'app/loadout-drawer/auto-loadouts';
 import { d2ManifestSelector, manifestSelector } from 'app/manifest/selectors';
+import { getCharacterProgressions } from 'app/progress/selectors';
 import { DimThunkDispatch, RootState, ThunkResult } from 'app/store/types';
 import { DimError } from 'app/utils/dim-error';
 import { errorLog, timer } from 'app/utils/log';
 import {
   DestinyCharacterComponent,
+  DestinyCharacterProgressionComponent,
   DestinyCollectibleComponent,
   DestinyCollectiblesComponent,
   DestinyComponentType,
@@ -35,6 +37,7 @@ import { showNotification } from '../notifications/notifications';
 import { loadingTracker } from '../shell/loading-tracker';
 import { reportException } from '../utils/exceptions';
 import { CharacterInfo, charactersUpdated, error, loadNewItems, update } from './actions';
+import { ArtifactXP } from './ArtifactXP';
 import { cleanInfos } from './dim-item-info';
 import { InventoryBuckets } from './inventory-buckets';
 import { DimItem } from './item-types';
@@ -270,6 +273,8 @@ export async function buildStores(
 
   const allItems = stores.flatMap((s) => s.items);
 
+  const characterProgress = getCharacterProgressions(profileInfo);
+
   const hasClassified = allItems.some(
     (i) =>
       i.classified &&
@@ -277,7 +282,7 @@ export async function buildStores(
   );
 
   for (const s of stores) {
-    updateBasePower(allItems, s, defs, hasClassified);
+    updateBasePower(allItems, s, defs, hasClassified, characterProgress);
   }
 
   return stores;
@@ -314,7 +319,7 @@ function processCharacter(
   const characterEquipment = profileInfo.characterEquipment.data?.[characterId]?.items || [];
   const itemComponents = profileInfo.itemComponents;
   const uninstancedItemObjectives =
-    profileInfo.characterProgressions?.data?.[characterId]?.uninstancedItemObjectives || [];
+    getCharacterProgressions(profileInfo, characterId)?.uninstancedItemObjectives || [];
 
   const store = makeCharacter(defs, character, lastPlayedDate);
 
@@ -402,7 +407,8 @@ function updateBasePower(
   allItems: DimItem[],
   store: DimStore,
   defs: D2ManifestDefinitions,
-  hasClassified: boolean
+  hasClassified: boolean,
+  characterProgress: DestinyCharacterProgressionComponent | undefined
 ) {
   if (!store.isVault) {
     const def = defs.Stat.get(StatHashes.Power);
@@ -432,6 +438,7 @@ function updateBasePower(
       name: t('Stats.PowerModifier'),
       hasClassified: false,
       description: '',
+      richTooltip: ArtifactXP(characterProgress),
       value: artifactPower,
       icon: xpIcon,
     };